[utils] extract split file support into helper file
Both update_verify_tests and update_generated_tests use the same utility
class for dealing with split-file tests. Extract it into a shared
helper.
